# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plymotion API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/system/themes/{dir_name}/frames/{index}": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Installed Frame",
        "operationId": "installed_frame_api_system_themes__dir_name__frames__index__get",
        "parameters": [
          {
            "name": "dir_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dir Name"
            }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/themes/{dir_name}/activate": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Activate",
        "operationId": "activate_api_system_themes__dir_name__activate_post",
        "parameters": [
          {
            "name": "dir_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dir Name"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/themes/{dir_name}": {
      "delete": {
        "tags": [
          "system"
        ],
        "summary": "Uninstall",
        "operationId": "uninstall_api_system_themes__dir_name__delete",
        "parameters": [
          {
            "name": "dir_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dir Name"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/restore-backup": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Restore Backup",
        "operationId": "restore_backup_api_system_restore_backup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreBackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/system/preview": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Preview",
        "operationId": "preview_api_system_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/login-logo/preview": {
      "post": {
        "tags": [
          "login-logo"
        ],
        "summary": "Make Preview",
        "operationId": "make_preview_api_login_logo_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login-logo/previews/{preview_id}": {
      "get": {
        "tags": [
          "login-logo"
        ],
        "summary": "Preview File",
        "operationId": "preview_file_api_login_logo_previews__preview_id__get",
        "parameters": [
          {
            "name": "preview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login-logo/apply": {
      "post": {
        "tags": [
          "login-logo"
        ],
        "summary": "Apply",
        "operationId": "apply_api_login_logo_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobInfo"
          }
        },
        "type": "object",
        "required": [
          "job"
        ],
        "title": "JobAccepted"
      },
...
      "JobInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "state": {
            "type": "string",
            "enum": [
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "State"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "result": {
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cancellable": {
            "type": "boolean",
            "title": "Cancellable"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "log_tail": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Log Tail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "state",
          "progress",
          "stage",
          "error",
          "cancellable",
          "created_at",
          "finished_at",
          "log_tail"
        ],
        "title": "JobInfo"
      },
...
      "LogoPreview": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "url",
          "width",
          "height"
        ],
        "title": "LogoPreview"
      },
      "LogoPreviewRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "max_height": {
            "type": "integer",
            "maximum": 400.0,
            "minimum": 16.0,
            "title": "Max Height",
            "default": 72
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "LogoPreviewRequest"
      },
...
      "PreviewRequest": {
        "properties": {
          "seconds": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 2.0,
            "title": "Seconds",
            "default": 6
          }
        },
        "type": "object",
        "title": "PreviewRequest"
      },
      "RestoreBackupRequest": {
        "properties": {
          "theme": {
            "type": "string",
            "title": "Theme"
          }
        },
        "type": "object",
        "required": [
          "theme"
        ],
        "title": "RestoreBackupRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}