# src/plymotion/api/routers/system.py
"""Themes installed on the system (/usr/share/plymouth/themes) and pkexec actions."""

from __future__ import annotations

from pathlib import Path

from fastapi import APIRouter, Request
from fastapi.responses import FileResponse

from plymotion.api.deps import SYSTEM_LOCK, ApiException, submit
from plymotion.api.schemas import InstalledTheme, JobAccepted, PreviewRequest, RestoreBackupRequest
from plymotion.core import installer, library
from plymotion.services import themes
from plymotion.services.progress import Reporter

router = APIRouter(prefix="/system", tags=["system"])


def _is_plymotion(theme: installer.InstalledTheme) -> bool:
    return "Generated by Plymotion" in theme.description


def _to_schema(theme: installer.InstalledTheme) -> InstalledTheme:
    frames = library.sorted_frames(theme.directory)
    base = f"/api/system/themes/{theme.directory.name}/frames"
    return InstalledTheme(
        dir_name=theme.directory.name, name=theme.name, description=theme.description,
        is_default=theme.is_default, is_plymotion=_is_plymotion(theme),
        has_backup=(installer.BACKUP_DIR / theme.directory.name).is_dir(),
        frame_count=len(frames),
        thumbnail_url=f"{base}/1" if frames else None,
        frame_url_template=f"{base}/{{n}}" if frames else None,
    )


def _theme_dir(dir_name: str) -> Path:
    try:
        return themes.installed_theme_dir(dir_name)
    except themes.NotFound as exc:
        raise ApiException(404, "theme_not_found", str(exc))


@router.get("/themes", response_model=list[InstalledTheme])
def list_installed() -> list[InstalledTheme]:
    items = installer.list_installed_themes()
    items.sort(key=lambda t: (not t.is_default, not _is_plymotion(t), t.name.lower()))
    return [_to_schema(t) for t in items]


@router.get("/themes/{dir_name}/frames/{index}", response_class=FileResponse)
def installed_frame(dir_name: str, index: int) -> FileResponse:
    try:
        path = themes.frame_path(_theme_dir(dir_name), index)
    except themes.NotFound as exc:
        raise ApiException(404, "frame_not_found", str(exc))
    return FileResponse(path, media_type="image/png", headers={"Cache-Control": "no-cache"})


@router.post("/themes/{dir_name}/activate", status_code=202, response_model=JobAccepted)
def activate(dir_name: str, request: Request) -> JobAccepted:
    name = _theme_dir(dir_name).name

    def run(reporter: Reporter) -> None:
        themes.run_system_action(
            f"Activando '{name}' con update-alternatives y regenerando el initramfs",
            lambda: installer.activate_theme(name), reporter,
        )

    return submit(request, "activate", f"Activar «{name}»", run, lock=SYSTEM_LOCK)


@router.delete("/themes/{dir_name}", status_code=202, response_model=JobAccepted)
def uninstall(dir_name: str, request: Request) -> JobAccepted:
    name = _theme_dir(dir_name).name
    if name == installer.TEXT_THEME_PLYMOUTH.parent.name:
        raise ApiException(400, "protected_theme", "El tema de texto es el respaldo seguro.")

    def run(reporter: Reporter) -> None:
        themes.run_system_action(
            f"Desinstalando '{name}' (si es el activo, se vuelve al tema de texto)",
            lambda: installer.uninstall_theme(name), reporter,
        )

    return submit(request, "uninstall", f"Desinstalar «{name}»", run, lock=SYSTEM_LOCK)


@router.post("/restore-backup", status_code=202, response_model=JobAccepted)
def restore_backup(body: RestoreBackupRequest, request: Request) -> JobAccepted:
    if not library.is_valid_slug(body.theme) or not (installer.BACKUP_DIR / body.theme).is_dir():
        raise ApiException(404, "no_backup", f"No hay backup de '{body.theme}'.")

    def run(reporter: Reporter) -> None:
        themes.run_system_action(
            f"Restaurando el backup de '{body.theme}' y regenerando el initramfs",
            lambda: installer.restore_backup(body.theme), reporter,
        )

    return submit(request, "restore-backup", f"Restaurar backup de «{body.theme}»", run,
                  lock=SYSTEM_LOCK)


@router.post("/reset-text", status_code=202, response_model=JobAccepted)
def reset_text(request: Request) -> JobAccepted:
    def run(reporter: Reporter) -> None:
        themes.run_system_action(
            "Volviendo al tema de texto y regenerando el initramfs",
            installer.reset_to_default, reporter,
        )

    return submit(request, "reset-text", "Volver a modo texto", run, lock=SYSTEM_LOCK)


@router.post("/preview", status_code=202, response_model=JobAccepted)
def preview(body: PreviewRequest, request: Request) -> JobAccepted:
    def run(reporter: Reporter) -> None:
        themes.run_system_action(
            f"Mostrando el tema activo durante {body.seconds} s (plymouthd)",
            lambda: installer.preview_installed_theme(seconds=body.seconds), reporter,
        )

    return submit(request, "preview", "Probar el tema activo", run, lock=SYSTEM_LOCK)